SqlGenerator - Move generateFieldSql back into the mixin class for backportability
